avoid overflowing width measure

+++ dmenu.c
@@ -89,10 +89,10 @@ calcoffsets(void)
 			break;
 }
 
-static int
+static long int
 max_textw(void)
 {
-	int len = 0;
+	long int len = 0;
 	for (struct item *item = items; item && item->text; item++)
 		len = MAX(TEXTW(item->text), len);
 	return len;
